+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 10.0.0 (unreleased)
 
-- Introduce alternative crypto backends (AWS-LC, RustCrypto)
+- BREAKING: use traits for crypto backends, you have to choose between `aws_lc_rs` and `rust_crypto`
+- Add `Clone` bound to `decode`
 
 ## 9.3.1 (2024-02-06)