+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - BREAKING: now using traits for crypto backends, you have to choose between `aws_lc_rs` and `rust_crypto`
 - Add `Clone` bound to `decode`
 - Support decoding byte slices
+- Support JWS
 
 ## 9.3.1 (2024-02-06)