+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 10.0.0 (unreleased)
+## 10.0.0 (2025-09-29)
 
 - BREAKING: now using traits for crypto backends, you have to choose between `aws_lc_rs` and `rust_crypto`
 - Add `Clone` bound to `decode`